refactor: extract logic to utils and clean up component structure
